+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 
 ### ⬆️ Improved
 - Add sending messages to only specific members. The `Message` entity contains a new property `restrictedVisibility` where a list of IDs for the desired members can be set.[#5644](https://github.com/GetStream/stream-chat-android/pull/5644)
+- Ensure that `TokenProvider.loadToken()` is not called multiple times while the token is being generated. [#5656](https://github.com/GetStream/stream-chat-android/pull/5656)
 
 ### ✅ Added